backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
Pillow==10.3.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
pdf2image==1.17.0
xgboost==2.0.3
shap==0.45.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
google-generativeai==0.5.4
supabase==2.4.3
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.3.0
aiofiles==23.2.1
python-jose==3.3.0
